Initial Branch Creation + Stateless
Added the NuGet Package Manager for Unity in order to get the Stateless package which is used for creating state machines easier. This created a couple of folders in the Assets folder, but I created a Libraries folder which holds the Stateless folder with the .dll and .xml files.
I also updated the packages-lock.json so that it no longer had the duplicated "com.unity.settings-manager" entry.

I'm including links to the GitHub pages for NuGetForUnity and Stateless for documentation and also a tutorial I was using to do it.

If we decide that we do not want to use Stateless for whatever reason then we can remove it.

Also created a new scripts folder for EnemyBehavior which will be used for the AI scripts.

https://www.gbrosgames.com/posts/Stateless-in-Unity/
https://github.com/GlitchEnzo/NuGetForUnity?tab=readme-ov-file
https://github.com/dotnet-state-machine/stateless